requirements-dev.txt
# Include production requirements
-r requirements.txt

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
factory-boy>=3.3.0
Faker>=20.1.0

# E2E testing
playwright>=1.40.0
pytest-playwright>=0.4.3

# Code quality
ruff>=0.1.6
black>=23.11.0
isort>=5.12.0

# Development tools
ipython>=8.18.1
django-debug-toolbar>=4.2.0
